Only reconnect on some error (Gateway & voice connections)
